backend_integrated/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database - MongoDB
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4

# Face Recognition - Multi-Model AI
insightface==0.7.3
deepface==0.0.79
dlib==19.24.2
opencv-python==4.8.1.78
face-recognition==1.3.0
mtcnn==0.1.1
onnxruntime==1.16.3

# Image Processing
Pillow==10.1.0
numpy==1.24.3
scikit-image==0.21.0
scipy==1.11.4

# Security & Authentication
bcrypt==4.1.1
cryptography==41.0.7
django-cors-headers==4.3.1
PyJWT==2.8.0

# Email & SMS Notifications
emails==0.6
twilio==8.10.0
premailer==3.10.0

# Reports & Analytics
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.3
matplotlib==3.8.2

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-redis==5.4.0

# Monitoring & Logging
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Development Tools
python-dotenv==1.0.0
ipython==8.18.1
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
